requirements.txt
click==6.3
colorama==0.3.6
python-firebase==1.2
requests==2.9.1
